+++ next.config.ts
@@ -13,6 +13,7 @@ const nextConfig: NextConfig = {
   images: {
     formats: ["image/avif", "image/webp"],
     minimumCacheTTL: 60 * 60 * 24 * 31,
+    qualities: [58, 62, 64, 68, 72],
     deviceSizes: [360, 390, 428, 640, 750, 828, 1080, 1200, 1920],
     imageSizes: [
       16, 32, 48, 64, 96, 128, 176, 200, 219, 242, 256, 300, 384,